Refactoring tracing
