+++ pkg/tracing/factory.go
@@ -2,45 +2,12 @@ package tracing
 
 import (
 	"context"
-	"github.com/improbable-eng/thanos/pkg/tracing/gcloud"
-	"github.com/improbable-eng/thanos/pkg/tracing/jaeger"
-	"gopkg.in/alecthomas/kingpin.v2"
 	"github.com/go-kit/kit/log"
-	"github.com/improbable-eng/thanos/pkg/tracing/factory"
 	"github.com/opentracing/opentracing-go"
+	"gopkg.in/alecthomas/kingpin.v2"
 )
 
-const (
-	jaegerTracingType = "jaeger"
-	gcloudTracingType = "gcloud"
-)
-
-// Factory - tracer factory.
-type Factory struct {
-	FactoryConfig
-	factories map[string]factory.Factory
-}
-
-// NewFactory return new tracer factory.
-func NewFactory(config FactoryConfig) (*Factory) {
-	f := &Factory{
-		FactoryConfig: config,
-	}
-	f.factories = make(map[string]factory.Factory)
-
-	f.factories[jaegerTracingType] = jaeger.NewFactory()
-	f.factories[gcloudTracingType] = gcloud.NewFactory()
-
-	return f
-}
-
-// Create implement factoty.Factory
-func (f *Factory) Create(ctx context.Context, logger log.Logger) (opentracing.Tracer, func() error) {
-	return f.factories[*f.FactoryConfig.TracingType].Create(ctx, logger)
-}
-
-func (f *Factory) RegisterKingpinFlags(app *kingpin.Application) {
-	for _, t := range f.factories {
-		t.RegisterKingpinFlags(app)
-	}
+type Factory interface {
+	Create(ctx context.Context, logger log.Logger) (opentracing.Tracer, func() error)
+	RegisterKingpinFlags(app *kingpin.Application)
 }
